Moved tests into AppDatabaseTest, reworked layout for battle and capture to avoid camera
