+++ app/src/androidTest/java/com/lutra/legallydistinctpocketmonsterarea/database/AppRepositoryTest.java
@@ -65,59 +65,4 @@ public class AppRepositoryTest extends TestCase {
     Log.d("getUserMonsterWithTypeById_Key", monsterEntry.getKey().toString());
     Log.d("getUserMonsterWithTypeById_Value", monsterEntry.getValue().toString());
   }
-
-  @Test
-  public void testInsertUserMonster() {
-    //Insert monster into DB, record monster ID
-    int userMonsterId = (int)repository.insertUserMonster(userMonster);
-
-    //Recall monster from DB to prove that it was inserted
-    UserMonster retrievedMonster = repository.getUserMonsterById(userMonsterId);
-    assertNotNull(retrievedMonster);
-
-    //Prove that it is identical to the monster we inserted
-    assertEquals(userMonster, retrievedMonster);
-
-    //Prove it is identical to a new monster we recall by the same ID
-    UserMonster retrievedMonster2 = repository.getUserMonsterById(userMonsterId);
-    assertEquals(retrievedMonster, retrievedMonster2);
-  }
-
-  @Test
-  public void testUpdateUserMonster() {
-    //Insert monster into DB, record ID
-    int userMonsterId = (int)repository.insertUserMonster(userMonster);
-
-    //Recall an instance of the monster by that ID, prove we were successful.
-    UserMonster beforeUpdate = repository.getUserMonsterById(userMonsterId);
-    assertNotNull(beforeUpdate);
-
-    //Prove the recalled monster does not have a given max health.
-    assertFalse(beforeUpdate.getMaxHealth() == NEW_HEALTH);
-
-    //Set the recalled monster's max health to a given value, and reinsert.
-    beforeUpdate.setMaxHealth(NEW_HEALTH);
-    repository.insertUserMonster(beforeUpdate);
-
-    //Use our original monster ID to recall the monster and prove that its max health has changed.
-    UserMonster afterUpdate = repository.getUserMonsterById(userMonsterId);
-    assertTrue(afterUpdate.getMaxHealth() == NEW_HEALTH);
-  }
-
-  @Test
-  public void testDeleteUserMonster() {
-    //Insert monster into DB, record ID
-    int userMonsterId = (int)repository.insertUserMonster(userMonster);
-
-    //Recall monster by ID and prove that it exists
-    UserMonster goodMonster = repository.getUserMonsterById(userMonsterId);
-    assertNotNull(goodMonster);
-
-    //Delete monster by ID from the database
-    repository.deleteMonsterByMonsterId(userMonsterId);
-
-    //Prove that we can't recall monster using that ID anymore
-    UserMonster badMonster = repository.getUserMonsterById(userMonsterId);
-    assertNull(badMonster);
-  }
 }
