+++ app/src/main/res/layout/activity_capture.xml
@@ -11,7 +11,7 @@
         android:id="@+id/enemyMonsterImage"
         android:layout_width="120dp"
         android:layout_height="120dp"
-        android:layout_marginTop="20dp"
+        android:layout_marginTop="80dp"
         android:layout_marginStart="20dp"
         android:contentDescription="@string/enemy_monster_image"
         app:layout_constraintStart_toStartOf="parent"
